Drag around one Image in the canvas.
It is very flickery. I don't see that there is any simple way to reduce
the flicker as long as we clear the whole canvas every time. One
solution might be to clear only the img rectangle that is being moved
but that will get complicated because the img could be overlapping other
ones, so we have to re-draw those (and only those). Do-able, but more
complicated than I want to deal with now.